Remove try/catch in tests, change tests folder structure and fix response statuses
