+++ package.json
@@ -8,7 +8,7 @@
     "api": "nodemon dist/server.js",
     "dev:api": "set NODE_ENV=dev&& concurrently \"npm run build\" \"npm run api\"",
     "test:api": "set NODE_ENV=test&& concurrently \"npm run build\" \"npm run api\"",
-    "db:init": "npm run db:migrate:all && npm run db:seed:all",
+    "db:init": "npm run db:create && db:migrate:all && npm run db:seed:undo && npm run db:seed:all",
     "db:create": "sequelize db:create --env=development",
     "db:migrate": "sequelize db:migrate --env=development",
     "db:migrate:all": "sequelize db:migrate --env=development",
@@ -16,7 +16,7 @@
     "movie:generator": "npm run build && node ./dist/scripts/movie-factory.js",
     "db:seed:all": "sequelize db:seed:all --env=development",
     "db:seed:undo": "sequelize db:seed:undo:all --env=development",
-    "tests": "set NODE_ENV=test&& mocha --require ts-node/register src/tests/**/*.ts"
+    "test": "set NODE_ENV=test&& mocha --require ts-node/register src/tests/**/*.ts"
   },
   "author": "Osvaldo Tejada",
   "license": "MIT",
@@ -42,8 +42,11 @@
     "@types/lodash": "^4.14.159",
     "@types/mocha": "^8.0.3",
     "@types/node": "^14.6.0",
+    "@typescript-eslint/eslint-plugin": "^3.10.1",
+    "@typescript-eslint/parser": "^3.10.1",
     "chai": "^4.2.0",
     "chai-http": "^4.3.0",
+    "eslint": "^7.7.0",
     "lorem-ipsum": "^2.0.3",
     "mocha": "^8.1.1",
     "nodemon": "^2.0.4",
